requirements.txt
asgiref==3.8.1
Django==5.1.3
django-ckeditor==6.7.1
django-environ==0.10.0
django-js-asset==2.2.0
djangorestframework==3.15.2
mysqlclient==2.2.4
pillow==11.0.0
psycopg2-binary==2.9.6
sqlparse==0.5.2
tzdata==2024.2
